# specs/explanation-framework/protocol/study.runtime-openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Invitations": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "default": 1,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "test": {
            "default": true,
            "title": "Test",
            "type": "boolean"
          }
        },
        "title": "Invitations",
        "type": "object"
      },
      "Mutation": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:/-]+$",
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "expected_revision"
        ],
        "title": "Mutation",
        "type": "object"
      },
...
      "Publish": {
        "additionalProperties": false,
        "properties": {
          "case_keys": {
            "items": {
              "$ref": "#/components/schemas/CaseKey"
            },
            "title": "Case Keys",
            "type": "array"
          },
          "definition": {
            "$ref": "#/components/schemas/StudyDefinition"
          }
        },
        "required": [
          "definition",
          "case_keys"
        ],
        "title": "Publish",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Exact ranking study",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/exports/{export_id}": {
      "get": {
        "operationId": "saved_export_api_v1_admin_exports__export_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Saved Export"
      }
    },
    "/api/v1/admin/invitations/{session_id}/reissue": {
      "post": {
        "operationId": "reissue_api_v1_admin_invitations__session_id__reissue_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reissue"
      }
    },
    "/api/v1/admin/invitations/{session_id}/revoke": {
      "post": {
        "operationId": "revoke_api_v1_admin_invitations__session_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke"
      }
    },
    "/api/v1/admin/studies": {
      "post": {
        "operationId": "publish_api_v1_admin_studies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publish"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish"
      }
    },
    "/api/v1/admin/studies/{revision}/close": {
      "post": {
        "operationId": "close_api_v1_admin_studies__revision__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close"
      }
    },
    "/api/v1/admin/studies/{revision}/exports": {
      "post": {
        "operationId": "export_api_v1_admin_studies__revision__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_test",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Test",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "include_keys",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Keys",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export"
      }
    },
    "/api/v1/admin/studies/{revision}/invitations": {
      "post": {
        "operationId": "issue_api_v1_admin_studies__revision__invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Invitations"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Issue"
      }
    },
    "/api/v1/admin/studies/{revision}/progress": {
      "get": {
        "operationId": "progress_api_v1_admin_studies__revision__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Progress"
      }
    },
...
    "/api/v1/study/cases/{case_id}/consultation": {
      "put": {
        "operationId": "consultation_api_v1_study_cases__case_id__consultation_put",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Expected session ID; a cookie switch returns 409 before mutation.",
            "in": "header",
            "name": "X-Study-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expected session ID; a cookie switch returns 409 before mutation.",
              "title": "X-Study-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consultation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consultation"
      }
    },
    "/api/v1/study/cases/{case_id}/draft": {
      "put": {
        "operationId": "draft_api_v1_study_cases__case_id__draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Expected session ID; a cookie switch returns 409 before mutation.",
            "in": "header",
            "name": "X-Study-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expected session ID; a cookie switch returns 409 before mutation.",
              "title": "X-Study-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ranking"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Draft"
      }
    },
    "/api/v1/study/cases/{case_id}/submit": {
      "post": {
        "operationId": "submit_api_v1_study_cases__case_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Expected session ID; a cookie switch returns 409 before mutation.",
            "in": "header",
            "name": "X-Study-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expected session ID; a cookie switch returns 409 before mutation.",
              "title": "X-Study-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Ranking"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit"
      }
    },
    "/api/v1/study/complete": {
      "post": {
        "operationId": "complete_api_v1_study_complete_post",
        "parameters": [
          {
            "description": "Expected session ID; a cookie switch returns 409 before mutation.",
            "in": "header",
            "name": "X-Study-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expected session ID; a cookie switch returns 409 before mutation.",
              "title": "X-Study-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mutation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete"
      }
    },
    "/api/v1/study/consent": {
      "put": {
        "operationId": "consent_api_v1_study_consent_put",
        "parameters": [
          {
            "description": "Expected session ID; a cookie switch returns 409 before mutation.",
            "in": "header",
            "name": "X-Study-Session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Expected session ID; a cookie switch returns 409 before mutation.",
              "title": "X-Study-Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Consent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consent"
      }
    },
...
  }
}